Added stay-after-save option to transaction form
A state variable `stayAfterSave` was added in Add component and passed through props to the Form component in transaction page. This adds a new feature allowing users to choose to stay on the page after saving a transaction, enhancing user experience. The Form was also updated to reset after submission if `stayAfterSave` option is enabled. A switch component was added in the UI for this new feature.