Refine transaction sort order in TransactionsGetListRequestHandler (#9)
Updated sorting of transactions to ensure deterministic ordering and prevent random sorting when two or more records have the same Date. This was done to ensure that when pagination occurs, no record is skipped. The sorting order now considers 'Date', 'Created', and 'Id' fields. Sorting by 'Id' is sufficient as it is unique but adding the 'Created' field also allows transactions to be sorted in the order they were entered.